Add suffix to modules with bug fixes
This patch add a suffix internal_{ansible_version} to all modules in
library. The {ansible_version} indicates which version of ansible
provides the fix, which we need to ship with the role.

So if Ansible {ansible_version} is released we can remove the module
from library and rename the task to the offical upstream module.

Related-to: https://bugzilla.redhat.com/show_bug.cgi?id=1487082